+++ models/get_cluster_settingses_request_body.go
@@ -13,6 +13,7 @@ import (
 )
 
 // GetClusterSettingsesRequestBody get cluster settingses request body
+// Example: {"after":"clusterSettingses-id-string","before":"clusterSettingses-id-string","first":0,"last":0,"orderBy":"createdAt_ASC","skip":0,"where":{"AND":"ClusterSettingsWhereInput[]","NOT":"ClusterSettingsWhereInput[]","OR":"ClusterSettingsWhereInput[]","cluster":"ClusterWhereInput","default_ha":false,"default_ha_not":false,"id":"string","id_contains":"string","id_ends_with":"string","id_gt":"string","id_gte":"string","id_in":["string"],"id_lt":"string","id_lte":"string","id_not":"string","id_not_contains":"string","id_not_ends_with":"string","id_not_in":["string"],"id_not_starts_with":"string","id_starts_with":"string"}}
 //
 // swagger:model GetClusterSettingsesRequestBody
 type GetClusterSettingsesRequestBody struct {
@@ -24,16 +25,16 @@ type GetClusterSettingsesRequestBody struct {
 	Before *string `json:"before,omitempty"`
 
 	// first
-	First *float64 `json:"first,omitempty"`
+	First *int32 `json:"first,omitempty"`
 
 	// last
-	Last *float64 `json:"last,omitempty"`
+	Last *int32 `json:"last,omitempty"`
 
 	// order by
 	OrderBy interface{} `json:"orderBy,omitempty"`
 
 	// skip
-	Skip *float64 `json:"skip,omitempty"`
+	Skip *int32 `json:"skip,omitempty"`
 
 	// where
 	Where interface{} `json:"where,omitempty"`
